Carry each beam option through to completion and repair scoring
We previously looked only two steps ahead for each beam option, but
that is insufficient for some. We now look ahead for all beam options
until we either find a new viable option or overrun. This means that
no option ends without having reached one of those conditions.

Score during alignment counts aligned and skipped equivalently, since
what it's really scoring is potential. Once the alignment is completed,
we care about aligned tokens and not about skipped ones. For that
reason, sort finished options to favor the ones that align the most
tokens (we were already doing that, although not reporting it correctly)
and break ties by favoring smallest number of blocks (which we were not
doing previously).